Updated add to cart
Fixed add to cart so that it reflects items on the page, and included removal functionality.
Streamlined fetch so that it doesn't need to delete and recreate from template, but instead attach a new cart item to the previous collection.
TODO:
add redundancies so that when quickly adding and removing cart items, the cart and page reflect the correct values.
Change initial fetch to check the localstorage and update quantities to reflect to the initial storefront.